requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
duckdb==1.2.1
greenlet==3.1.1
idna==3.10
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
pandas==2.2.3
Pygments==2.19.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rich==14.0.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
win32_setctime==1.2.0
